+++ src/sgpp/base/grid/generation/functors/SurplusVolumeRefinementFunctor.hpp
@@ -53,7 +53,7 @@ protected:
 	DataVector* alpha;
 
 	/// number of grid points to refine
-	int refinements_num;
+	size_t refinements_num;
 
 	/// threshold, only the points with greater to equal absolute values of the refinement criterion (e.g. alpha or error) will be refined
 	double threshold;
